Refactor types files
