+++ src/config/passport.ts
@@ -1,6 +1,6 @@
 import multiStrategy from './strategies/multi-saml-strategy';
 import { PassportStatic } from 'passport';
-import { PassAuthAppConfig } from '../model/model';
+import { PassAuthAppConfig } from '../@types';
 
 export default function (passport: PassportStatic, config: PassAuthAppConfig) {
   passport.serializeUser(function (user, cb) {
